Remove the monkeypatching module entirely.
This has been the most brittle part of the library, and not really
worth it just for some syntactic sugar (that doesn't even work with
lsp/autocomplete), as nice as having access to (array).sqrt() and so
on would be.

Also, cleanup and fix a bunch of linter complaints, and remove
deprecated Module functions.